refactor: move class
여러 테스트에서 사용하는 테스팅 객체를 상위 패키지로 이동
